fix: cast enum bind flag to UINT in NV12 converter ternary
CI's MinGW GNU GCC compiles the native target with
-Wall -Wextra -Wpedantic -Werror, which rejects a ternary whose two
arms are an enumerated type (D3D11_BIND_VIDEO_ENCODER) and a plain
unsigned literal (0u) as "enumerated and non-enumerated type in
conditional expression [-Werror=extra]". Local/MSVC (/W4 /WX) and
clang-mingw do not flag it, so it was latent until the manifest gate
stopped masking the native build. Cast the enum to UINT so both arms
share the non-enumerated type BindFlags already expects.

Verified with GNU GCC 16 -O3 -c across all 11 native translation
units (the 5 CI never reached included) using CI's exact flag set.

+++ src/client/native/src/gpu_color_converter.cpp
@@ -57,7 +57,7 @@ HRESULT ensure_converter(rs_encoder_t* encoder, const D3D11_TEXTURE2D_DESC& sour
   output.SampleDesc.Count = 1;
   output.Usage = D3D11_USAGE_DEFAULT;
   output.BindFlags = D3D11_BIND_RENDER_TARGET |
-      (needs_video_encoder_binding ? D3D11_BIND_VIDEO_ENCODER : 0u);
+      (needs_video_encoder_binding ? static_cast<UINT>(D3D11_BIND_VIDEO_ENCODER) : 0u);
   if (SUCCEEDED(result)) result = encoder->runtime->device->CreateTexture2D(&output, nullptr, &encoder->nv12_surface);
   if (FAILED(result)) return result;
